pyproject.toml
[project]
name = "queryscaler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.5.1",
    "faker>=40.13.0",
    "openenv-core[core]>=0.2.3",
    "sqlglot>=30.3.0",
]

[project.scripts]
server = "server.app:main"
